refactor(test): P5-1 continuation — extract canvas2d tests from test_widget_bridge.cpp
Phase 5 P5-1 continuation (second slice after runtime-import in
PR #2160). Pull the 1,224-line Canvas2D-bridge cluster out of
test_widget_bridge.cpp into a sibling test_widget_bridge_canvas2d.cpp
TU, covering three adjacent concerns:

- canvasSetTransform / canvasClip / canvasGlobalCompositeOperation
  (issue-896)
- canvasMeasureText / canvasSetLineDash / canvasDrawImage /
  canvasGetImageData / canvasPutImageData (issue-916)
- pulp #1899 / #1901 — 4-arg canvasFillText preserves prior state

- test/test_widget_bridge.cpp: 13,609 → 12,385 lines (-1,224 net).
- test/test_widget_bridge_canvas2d.cpp: 1,281 lines (new).
- New CMake target pulp-test-widget-bridge-canvas2d via
  pulp_add_test_suite() helper from R2-7.

The `canvasFromBridge` helper that was originally inside the
extracted block (anonymous namespace) had to be kept in the parent
TU too: 10+ other tests (CSS clip-path, box-shadow JS bridge, SVG
widget integration, etc.) call it. Each TU now has its own copy in
its anonymous namespace; a future shared
test/test_widget_bridge_helpers.hpp could centralise it.

Validation:
- pulp-test-widget-bridge (parent, after extraction): 358 cases /
  2,057 assertions; 3 cases / 4 assertions fail with pre-existing
  font_family selection regressions at lines 6978/10707/12354/12393
  (unrelated to this split — same failures on origin/main per the
  PR #2169 font-v2 hardening that recently landed).
- pulp-test-widget-bridge-canvas2d (new): 26 cases / 196 assertions —
  all pass.

P5-1 progress: runtime-import (PR #2160) + canvas2d (this PR).
Remaining surfaces: input, dom, CSS, RN-style, SVG, html — each can
ship in its own follow-up PR.
